+++ core-service/src/main/java/com/geekbrains/ru/gb_rest_angular/core/dto/OrderDto.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -15,7 +16,7 @@ public class OrderDto {
     public Long id;
     private Long userId;
     private List<OrderItem> items;
-    public Integer cost;
+    public BigDecimal cost;
     private String address;
     private String phone;
