+++ core-service/src/main/resources/db/migration/V4__addOrders.sql
@@ -1,7 +1,7 @@
 CREATE TABLE orders (
                         id bigserial primary key,
                         user_id bigint not null references users(id),
-                        cost int not null,
+                        cost numeric(8,2) not null,
                         address varchar(255),
                         phone varchar(255),
                         created_at timestamp default current_timestamp,
@@ -14,8 +14,8 @@ CREATE TABLE  order_items (
                              product_id bigint not null references products(id),
                              order_id bigint not null references orders(id),
                              quantity int not null,
-                             price_per_product int not null,
-                             cost int not null,
+                             price_per_product numeric(8,2) not null,
+                             cost numeric(8,2) not null,
                              created_at timestamp default current_timestamp,
                              updated_at timestamp default current_timestamp
 );
